+++ FirebaseCombineSwift/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - [feature] Added Combine support for Cloud Functions for Firebase
 - [feature] WIP: Added Combine support for Firebase Auth
+- [feature] WIP: Added Combine support for Firebase Storage